editedPlayerMovement
edited PlayerMovement. Player couldn't move since the scene was always getting reloaded since the position of the player at the start was always smaller than 10 so the scene resatarted a infinite number of times and you could never move.
also i implemented, that the lives count down on the top of the UI.