new scene and logic
added a new scene that you can access via portal. added a gun pickup in the next scene and that you can move it with your mouse. doesn't work quite yet. you lose all the abilities you had in the last scvene. that's because fo dontdestroyonload();
UI is also gone in the next scene but that doesn't matter.